% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metades.R
\name{metades}
\alias{metades}
\title{metades}
\usage{
data(gene_data)

data(pheno)

data(symbol)

res<-metagenescore(gene_data,symbol,method="zscore")

metades_results<-imdes(res,pheno$group,control="Non_Tumor")
}
\arguments{
\item{group}{group column from phenotype file}

\item{control}{string defining control in pheno$group data}
}
\description{
metades
}
\examples{
data(gene_data)
data(pheno)
data(symbol)
res<-metagenescore(data,symbol,method="zscore")
metades_results<-metades(res,pheno$group,control="Non_Tumor")
head(metades_results)
}
