% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaensemblscore.r
\name{metaensemblscore}
\alias{metaensemblscore}
\title{metaensemblscore}
\usage{
data(ensembl_data)

data(pheno)

data(ensembl)

res<-metaensemblscore(ensembl_data,ensembl,method="zscore")

head(res[,1:6])
}
\arguments{
\item{ensembl}{metabolism modulation genesets by ensembl identifiers}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}

\item{data}{transcriptome matrix annotated with ensembl identifiers}
}
\description{
metaensemblscore
}
\examples{
data(ensembl_data)
data(pheno)
data(ensembl)
res<-metaensemblscore(ensembl_data,ensembl,method="zscore")
head(res[,1:6])
}
