% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenescore.R
\name{metagenescore}
\alias{metagenescore}
\title{metagenescore}
\usage{
data(gene_data)

data(pheno)

data(symbol)

res<-metagenescore(gene_data,symbol,method="zscore")

head(res[,1:6])
}
\arguments{
\item{symbol}{kegg metabolism genesets by gene symbols}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}

\item{data}{transcriptome matrix}
}
\description{
metagenescore
}
\examples{
data(gene_data)
data(pheno)
data(symbol)
res<-metagenescore(gene_data,symbol,method="zscore")
head(res[,1:6])
}
