% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaheat.r
\name{metaheat}
\alias{metaheat}
\title{metaheat}
\usage{
data(gene_data)

data(symbol)

data(pheno)

res<-metagenescore(gene_data,symbol,method="zscore")

metaheat(res,pheno,scale="none",fontsize=10)
}
\arguments{
\item{results}{results obtained from metagenescore function or sub-geneset}

\item{pheno}{phenotype file from tested samples}

\item{scale}{method of scaling "none" "row" "column"}

\item{fontsize}{size of the font on the graph}
}
\description{
metaheat
}
\examples{
data(gene_data)
data(symbol)
data(pheno)
res<-metagenescore(gene_data,symbol,method="zscore")
metaheat(res,pheno,scale="none",fontsize=10)
}
