% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaprotscore.R
\name{metaprotscore}
\alias{metaprotscore}
\title{metaprotscore}
\usage{
data(swissprot_data)

data(pheno)

data(swissprot)

res<-metaprotscore(swissprot_data,swissprot,method="zscore")

head(res[,1:6])
}
\arguments{
\item{swissprot}{kegg metabolism genesets by swissprot id}

\item{method}{GSVA method of enrichment: "gsva", "ssgsea", "zscore", "plage"}

\item{data}{transcriptome matrix}
}
\description{
metaprotscore
}
\examples{
data(swissprot_data)
data(pheno)
data(swissprot)
res<-metaprotscore(swissprot_data,swissprot,method="zscore")
head(res[,1:6])
}
