% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metavol.R
\name{metavol}
\alias{metavol}
\title{metavol}
\usage{
data(gene_data)

data(symbol)

data(pheno)

res<-metagenescore(gene_data,symbol,method="zscore")

metades_results<-metades(res,pheno$group,control="Non_Tumor")

metavol(metades_results,label=3.5,font=12,title="KEGG metabolism")
}
\arguments{
\item{res}{results obtained from metades function}

\item{label}{font size for geneset labels}

\item{font}{font size for titles}

\item{title}{title description}
}
\description{
metavol
}
\examples{
data(gene_data)
data(symbol)
data(pheno)
res<-metagenescore(gene_data,symbol,method="zscore")
metades_results<-metades(res,pheno$group,control="Non_Tumor")
metavol(metades_results,label=3.5,font=12,title="KEGG metabolism")
}
